requirements.txt
cryptography
python-dotenv
watchdog
pywin32
reportlab
pyinstaller